backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for finance chatbot"
authors = [
    {name = "Michael Hua",email = "michael.huayl@yahoo.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.14"
langchain-openai = "^0.3.0"
langgraph = "^0.2.62"
langchain-community = "^0.3.14"
pymupdf = "^1.25.1"
python-pptx = "^1.0.2"
pandas = "^2.2.3"
sqlalchemy = "^2.0.37"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
fastapi = {extras = ["standard"], version = "^0.115.6"}
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
chromadb-client = "^0.6.3"
nanoid = "^2.0.0"
langchain-chroma = "^0.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
pytest = "^8.3.4"

[tool.ruff]
show-fixes = true
target-version = "py312"
line-length = 80

[tool.ruff.lint]
extend-select = ["N", "UP", "PL", "Q"]
ignore = ["E501"]

